Fix socket mode on connect and login, change code style
"`connect/2` sets the socket to `active: :once` on `:gen_tcp.connect/4`
and then again with `:inet.setopts/2`. This could result in two tcp
messages. Since `login/2` occurs immediately afterwards it would be
better to set the socket to `active: false` on connect and not set the
socket to active until after login has succeeded." - @fishcakez